+++ CHANGELOG.md
@@ -2,3 +2,5 @@ CHANGELOG.md
 + version 1.3.0 - v1.4.0 on experiment
 + version 1.4.0 for-key bug fixing
 + version 1.4.4 added onSuspense and fix out- (document event)
++ version 1.4.7 - 1.4.10 solving resuming issue when wrapping with template
++ version 1.4.11 - added more event modifiers